requirements.txt
duckduckgo-search
googlesearch-python
